backend/requirements.txt
google-genai>=1.0.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
sse-starlette>=2.0.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
